Make headers imported in umbrella header public
The headers that are included in `TrustKit.h` need to be made public or
the framework will generate warnings when including.

This also removes the custom module map. Xcode will automatically do
the right thing and use `TrustKit.h`. This makes maintainance easier
because the contents of `TrustKit.h` don't have to be duplicated in
`TrustKit-umbrella.h` as well.