Fix domain selection in overlapping pinsets
Observed behavior of TrustKit showed that, when a domain did not
have an exact match for a pinset, the first matching pinset config
with the IncludeSubdomains flag was selected. This led to
unpredictable behavior because of the nature of iterating through
dictionary keys. The change in this commit modifies TrustKit's
selection algorithm to iterate through all pinset configs and then
select the one that is the closest match (e.g, longest domain).
This matches the industry best practice set by Google in their
native Android pinning implementation, and brings TrustKit's
behavior in line with that of TrustKit-Android.